refactor(pay-card): keep every login shape in types.ts
`ports.ts` and `types.ts` each imported types from the other: the ports are
built from the flow's own shapes, and the machine context holds the ports. The
cycle was harmless, because both files were type-only and every import was
`import type`, but a reader had to hold two files open to follow one contract.

One file now, in four sections: the attempt and the callback, the ports, the
machine's own shapes, and the Redux slice state. Nothing is renamed, so the
package's public API is unchanged.
