+++ features/flow/pay-card-auth/src/index.native.ts
@@ -1,4 +1,4 @@
 export { CardLogin } from "./components/CardLogin/index.native";
 export type { CardLoginProps } from "./components/CardLogin/types";
-export type { OpenHostedLogin, HostedLoginResult } from "./state/ports";
+export type { OpenHostedLogin, HostedLoginResult } from "./state/types";
 export type { CardLoginOauthConfig, PayCardAuthCallback } from "./state/types";
